+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - [DX] `Database.get(tableName)` is now a shortcut for `Database.collections.get(tableName)`
 - [DX] Query is now thenable - you can now use `await query` and `await query.count` instead of `await query.fetch()` and `await query.fetchCount()`
 - [DX] Relation is now thenable - you can now use `await relation` instead of `await relation.fetch()`
+- [DX] Exposed `collection.db` and `model.db` as shortcuts to get to their Database object
 
 ### Changes